Add scan tests
